Add node exception file
This file will be a json list of all node IDs that have been tagged
but aren't technically wrong as mapped. This will prevent them from
showing up in future maps that are told to ignore IDs in this file.
The file need not be sorted. Just be careful that there isn't a comma
on the final value since json doesn't allow that.